service/backend/requirements.txt
flask
flask-pymongo
gunicorn
pyjwt
werkzeug
pycryptodome
